broke the int32 into several data structures in the base voxel grid

+++ Assets/VoxelPainter/Rendering/MarchingCubes/IMarchingCubesVisualizer.cs
@@ -24,6 +24,6 @@ namespace VoxelPainter.VoxelVisualization
         /// </summary>
         /// <param name="verticesValues"></param>
         /// <param name="vertexAmount"></param>
-        void GetVerticesValuesNative(ref NativeArray<float> verticesValues, Vector3Int vertexAmount);
+        void GetVerticesValuesNative(ref NativeArray<int> verticesValues, Vector3Int vertexAmount);
     }
 }
